Add files via upload
Online Reservations System: 
This Online Reservation System will be easy to use and can be used by any person. The basic idea behind this project is to save data in a central database which can be
accessed by any authorize person to get information and saves time and burden which are being
faced by their customers.